requirements.txt
click==8.1.3
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.21.0
SQLAlchemy==1.4.42
Werkzeug==2.2.2
zipp==3.10.0
